+++ src/index.js
@@ -3,10 +3,8 @@ import standalone from "./standalone"
 import createPlugin from "./createPlugin"
 import rules from "./rules"
 import {
-  cssWordIsVariable, // Deprecated
   report,
   ruleMessages,
-  styleSearch,
   validateOptions,
 } from "./utils"
 import createRuleTester from "./testUtils/createRuleTester"
@@ -14,10 +12,8 @@ import createRuleTester from "./testUtils/createRuleTester"
 const stylelint = postcssPlugin
 
 stylelint.utils = {
-  cssWordIsVariable, // Deprecated
   report,
   ruleMessages,
-  styleSearch,
   validateOptions,
 }