+++ src/rules/selector-pseudo-class-parentheses-space-inside/index.js
@@ -1,10 +1,10 @@
+import styleSearch from "style-search"
 import _ from "lodash"
 import {
   isStandardSyntaxRule,
   parseSelector,
   report,
   ruleMessages,
-  styleSearch,
   validateOptions,
 } from "../../utils"